+++ app/Http/Controllers/Quest/QuestCommentController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Http\Controllers\Quest;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+
+class QuestCommentController extends Controller
+{
+    //
+}
