+++ app/Http/Controllers/Quest/QuestLikeController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Business;
+namespace App\Http\Controllers\Quest;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
