+++ app/Http/Controllers/Spot/SpotController.php
