+++ app/Http/Controllers/Spot/SpotLikeController.php
@@ -8,7 +8,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
-class LikeController extends Controller
+class SpotLikeController extends Controller
 {
     private $like;
