+++ app/Models/QuestBody.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class QuestBody extends Model
+{
+    protected $table = 'quest_bodies';
+}
