+++ app/Models/QuestComment.php
@@ -8,6 +8,7 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 class QuestComment extends Model
 {
     use SoftDeletes;
+    protected $table = 'quest_comments';
 
     public function quest(){
         return $this->belongsTo(Quest::class);
