+++ app/Models/QuestCommentLike.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class QuestCommentLike extends Model
+{
+    protected $table = 'quest_comment_likes';
+}
